Update: Delete portfolio/{id}/block api 스펙 변경
- HTTP Status 204를 반환하도록 변경
- validation 로직을 Controller로 위임

+++ src/main/java/com/jo2seo/aomd/portfolio/PortfolioController.java
@@ -82,10 +82,13 @@ public class PortfolioController {
     }
 
     @DeleteMapping("/portfolio/{id}/block")
-    public void deletePortfolioBlock(
+    public ResponseEntity deletePortfolioBlock(
             @PathVariable("id") String shareUrl,
             @Valid @RequestBody DeletePortfolioBlockRequest deletePortfolioBlockRequest
-    ) throws BaseException {
+    ) {
+        boolean isMine = portfolioService.checkMine(shareUrl);
+        if (!isMine) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         portfolioService.deleteBlock(shareUrl, deletePortfolioBlockRequest.getBlockId());
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 }
